requirements.txt
# Core frameworks
streamlit

# LLM/RAG tooling
langchain
llama-index
llama-index-embeddings-huggingface
llama-index-llms-huggingface
llama-index-vector-stores-faiss
sentence-transformers
transformers
accelerate
einops
bitsandbytes

# Vector Store
faiss-cpu

# File handling
python-docx
pypdf

# Utilities
tiktoken
docx2txt
docxcompose

python-dotenv



